Reimplement spread mode without precomputing values
Previously spread mode was implemented by applying an offset to the
data values and storing them in a separate vector.  ImPlot now allows
you to specify a custom function, which is used to apply the spread
mode transformation as the plot is drawn.

This uses half as much memory, and it should speed up startup.  Frame
rate is probably lower now, but it hasn't been noticeable so far.

It also allows vertical zoom to be applied in spread plot mode.